requirements.txt
tqdm==4.19.9
numpy==1.23.5
tensorflow==2.15.0
tensorboardX
torch==2.4.0
Pillow==6.2.0
torchvision==0.2.2


